move date function to utils and del hardcode de-DE
